PS-9040 Integrate clang-tidy checks for Percona server
https://perconadev.atlassian.net/browse/PS-9040

Creates a yaml file for clang-tidy workflow which is triggered
on a pull request.

The workflow triggers a run of generating comiplation commands
and uses the same to perform a clang-tidy check on the modified
files and generates a report of the warnings in the changed code
through a github action bot in the form of review comments on
the pull request.

Disable some directories to run currus CI and azure pipeline
changes for testing.